Create Format class to handle telefone number formatting
