+++ TesteiOS/TesteiOS/Objects/Format.swift
@@ -1,0 +1,52 @@
+//
+//  NumberFormat.swift
+//  TesteiOS
+//
+//  Created by Carolina Bonturi on 7/5/18.
+//  Copyright © 2018 Carolina Bonturi. All rights reserved.
+//
+
+import Foundation
+
+
+class Format {
+    
+    public static let shared = Format()
+    
+    func format(phoneNumber: String, shouldRemoveLastDigit: Bool = false) -> String {
+        
+        guard !phoneNumber.isEmpty else { return "" }
+        guard let regex = try? NSRegularExpression(pattern: "[\\s-\\(\\)]", options: .caseInsensitive) else { return "" }
+        let r = NSString(string: phoneNumber).range(of: phoneNumber)
+        var number = regex.stringByReplacingMatches(in: phoneNumber, options: .init(rawValue: 0), range: r, withTemplate: "")
+        
+        // max 11 numbers
+        if number.count > 11 {
+            let eleventhDigitIndex = number.index(number.startIndex, offsetBy: 11)
+            number = String(number[number.startIndex..<eleventhDigitIndex])
+        }
+        
+        if shouldRemoveLastDigit {
+            let end = number.index(number.startIndex, offsetBy: number.count-1)
+            number = String(number[number.startIndex..<end])
+        }
+        
+        if number.count < 7 {
+            let end = number.index(number.startIndex, offsetBy: number.count)
+            let range = number.startIndex..<end
+            number = number.replacingOccurrences(of: "(\\d{2})(\\d+)", with: "($1) $2", options: .regularExpression, range: range)
+            
+        } else if number.count < 11 {
+            let end = number.index(number.startIndex, offsetBy: number.count)
+            let range = number.startIndex..<end
+            number = number.replacingOccurrences(of: "(\\d{2})(\\d{4})(\\d+)", with: "($1) $2-$3", options: .regularExpression, range: range)
+        } else {
+            let end = number.index(number.startIndex, offsetBy: number.count)
+            let range = number.startIndex..<end
+            number = number.replacingOccurrences(of: "(\\d{2})(\\d{5})(\\d+)", with: "($1) $2-$3", options: .regularExpression, range: range)
+        }
+        
+        return number
+    }
+    
+}
